Bug 1580087. Fix 32bit * 32bit integer multiplies so they don't overflow in imagelib. r=aosmond
Differential Revision: https://phabricator.services.mozilla.com/D45303

+++ image/SurfaceCache.cpp
@@ -1644,7 +1644,7 @@ IntSize SurfaceCache::ClampVectorSize(const IntSize& aSize) {
     return aSize;
   }
 
-  int32_t proposedKB = int32_t(int64_t(aSize.width) * aSize.height / 256);
+  int64_t proposedKB = int64_t(aSize.width) * aSize.height / 256;
   if (maxSizeKB >= proposedKB) {
     return aSize;
   }
